Infer input types
Trying to pop a value off an empty stack implies that a value should
have been on the stack to begin with.  This lets us infer a program's
inputs.  Since a single operator is a program, it also lets us infer
the type signature of an operator.

For example, ghci gives the following behavior:

    > run (parse "+") (TypeMachine [] [])
    TypeMachine {typeStack = [TypeInt], typeQueue = [TypeInt,TypeInt]}

The input types are shown in reverse order.