Implement code value execution
For the DataMachine, executing code is just copying the code's ops to
the code queue.  The next iteration will start running the ops.  This
gives the DataMachine power to loop forever.  For example, `[d!]d!`.

For the TypeMachine, we can't copy code to the code queue.  That
allows infinite recursion which means the type checker may never
finish.  Fortunately, when the code value was pushed onto the stack,
we inferred its type.  That means we already know what values it will
consume and produce.  Consuming and producing those values performs
the type check while still guaranteeing progress through the code
queue.

The type check currently fails to check `[d!]d!`.  It'd be cool if it
caught the infinite loop for me, but it's really a fault in the type
checker.  It doesn't handle unknown types correctly.  TypeUnknown
should be replaced with TypeVar Int representing a specific unknown
type.  That way, once we discover what that specific type should be,
we can update the machine to reflect it.