First (failed) attempt at resolving type variables
The goal is to correctly infer types for scripts like `d+` and `d,`
(where `d` creates type variables because it's polymorphic).  This
code correctly infers the type of `d+` but leaves a floating type
variable for `d,`.

The problem is in the op implementations.  OpAppend is defined in such
a way that the first type variable on the stack is removed from the
machine before a type variable is resolved.  Resolving the variable by
replacing all type variables in the machine can't access the one that
escaped the machine.  When that type variable is added back into the
machine (with `push`), it's still a type variable.

The type variable resolutions must somehow be accumulated and applied
once we know that all values are back in the machine.